added tau daughters pdg + charge vectors

+++ source/include/eventVariables.h
@@ -41,7 +41,7 @@ class eventVariables{
 	std::vector<int> _MCfpdg{};
 	std::vector<TLorentzVector> _MCTauVisibleDaughters{};
 	std::vector<int> _MCTauVisibleDaughters_pdg{};
-	std::vector<int> _MCTauVisibleDaughters_charge{};
+	std::vector<double> _MCTauVisibleDaughters_charge{};
 
 	TLorentzVector* _mcl{};
 	TLorentzVector* _mcqq{};
